requirements.txt
numpy
spacy
nltk
syllapy
pyspellchecker
textstat